Planet gravity tuning

+++ WrathOfKhan/Assets/Scenes/Gameplay/Planet/GravBodyScript.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class GravBodyScript : MonoBehaviour {
 
-    public float GM;
+    public double GM;
     public float radius { get { return GetComponent<CircleCollider2D>().radius; } }
 
     // Use this for initialization
